requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
importlib-metadata==8.4.0
Markdown==3.7
pillow==10.4.0
sqlparse==0.5.1
typing-extensions==4.12.2
zipp==3.20.1
